refactoring: reducer 모듈을 reducer 폴더 로 이동
