+++ frontend/guest-app/src/components/Question/QuestionsContext.js
@@ -6,7 +6,7 @@ import {
 	QUERY_INIT_QUESTIONS,
 } from "../../libs/useQueryQuestions.js";
 import {useSocket} from "../../libs/socketIoClientProvider.js";
-import QuestionsRepliesReducer from "./QuestionsRepliesReducer.js";
+import QuestionsRepliesReducer from "../../reducer/QuestionsRepliesReducer.js";
 
 const QuestionsContext = createContext([]);
