+++ frontend/guest-app/src/components/UIController/UIControllerProvider.js
@@ -1,5 +1,5 @@
 import React, {createContext, useContext} from "react";
-import useToggleReducer from "../Question/useToggleReducer.js";
+import useToggleReducer from "../../reducer/useToggleReducer.js";
 
 const UIControlContext = createContext([]);
