+++ frontend/guest-app/src/reducer/ToggleReducer.js
